Change release to use xz for tar archives
This changes the release to use .tar.xz instead of .tar.gz archives for
the HTML book and source archive, which reduces the archive size by a
factor of 5 to 10. The QCH book doesn't really benefit from using xz but
takes considerably longer to compress, so I'll leave this one.